+++ src/core/serviceContainer.ts
@@ -20,6 +20,7 @@ import { RepositoryManager } from "../features/ai-template-files/services/reposi
 import { ProfileService } from "../features/profile-management/services/profileService";
 import { ModeSelectionService } from "../features/initialization/modeSelectionService";
 import { DevOpsMcpConfigService } from "../features/apm-devops/devOpsMcpConfigService";
+import { NexkitFileMigrationService } from "../features/initialization/nexkitFileMigrationService";
 import { CommitMessageService } from "../features/commit-management/commitMessageService";
 
 /**
@@ -47,6 +48,7 @@ export interface ServiceContainer {
   profileService: ProfileService;
   modeSelection: ModeSelectionService;
   devOpsConfig: DevOpsMcpConfigService;
+  nexkitFileMigration: NexkitFileMigrationService;
   commitMessage: CommitMessageService;
 }
 
@@ -84,6 +86,7 @@ export async function initializeServices(context: vscode.ExtensionContext): Prom
   const profileService = new ProfileService(installedTemplatesState, aiTemplateData, backup);
   const modeSelection = new ModeSelectionService();
   const devOpsConfig = new DevOpsMcpConfigService();
+  const nexkitFileMigration = new NexkitFileMigrationService();
   const commitMessage = new CommitMessageService();
 
   // Register for disposal
@@ -115,6 +118,7 @@ export async function initializeServices(context: vscode.ExtensionContext): Prom
     profileService,
     modeSelection,
     devOpsConfig,
+    nexkitFileMigration,
     commitMessage,
   };
 }