+++ CHANGELOG.md
@@ -1,3 +1,17 @@
+## 3.5.0-beta.2 (2026-02-26)
+
+* Merge branch 'develop' of https://github.com/NexusInnovation/nexus-nexkit-vscode into develop ([50a851a](https://github.com/NexusInnovation/nexus-nexkit-vscode/commit/50a851a))
+* feat: Include Full BMAD Method ([cf1cab3](https://github.com/NexusInnovation/nexus-nexkit-vscode/commit/cf1cab3))
+
+## 3.5.0-beta.1 (2026-02-26)
+
+* Include BMAD Method ([fb52746](https://github.com/NexusInnovation/nexus-nexkit-vscode/commit/fb52746))
+* Merge branch 'main' into develop ([9a9b424](https://github.com/NexusInnovation/nexus-nexkit-vscode/commit/9a9b424))
+* chore(release): v3.4.0-beta.1 [skip ci] ([9c8c1f3](https://github.com/NexusInnovation/nexus-nexkit-vscode/commit/9c8c1f3))
+* fix: Merge pull request #93 from NexusInnovation/feature/move-nexkit-files-to-.nexkit ([9b12d84](https://github.com/NexusInnovation/nexus-nexkit-vscode/commit/9b12d84)), closes [#93](https://github.com/NexusInnovation/nexus-nexkit-vscode/issues/93)
+* feat: move Nexkit template files from .github to .nexkit directory ([416c842](https://github.com/NexusInnovation/nexus-nexkit-vscode/commit/416c842))
+* feat(initialization): add Nexkit file migration service ([71fed35](https://github.com/NexusInnovation/nexus-nexkit-vscode/commit/71fed35))
+
 ## 3.4.0 (2026-02-26)
 
 * Initial plan ([886cc10](https://github.com/NexusInnovation/nexus-nexkit-vscode/commit/886cc10))